# Update sampling interval back to 5ms