@rmccue 2023-01-31
> There's no notable change on the performance.

Maybe I'm missing something, but that's not how I'm reading the results; average response time went up from 0.87s to 1.19s, an increase of 0.32s/36%? That seems a pretty big difference though.