@jerico 2023-01-31
> > There's no notable change on the performance.
> 
> 
> 
> Maybe I'm missing something, but that's not how I'm reading the results; average response time went up from 0.87s to 1.19s, an increase of 0.32s/36%? That seems a pretty big difference though.

Sorry I averaged this in my head as still "around 1 second". That's where my comment came from. 

Should I also do a test on how the original xhprof was performing? How should we know the threshold of acceptable range?