@jerico 2023-02-01
@rmccue I noticed that when I have another running task, it affects the response time. 

I went ahead an re-ran the test. This time, more conscious of what other processes are running. For the following tests, I only have 1 instance of Altis project running, and no other apps aside from macOS system services running.

```
## 50ms

Transactions:		         500 hits
Availability:		      100.00 %
Elapsed time:		      248.56 secs
Data transferred:	        3.68 MB
Response time:		        0.50 secs
Transaction rate:	        2.01 trans/sec
Throughput:		        0.01 MB/sec
Concurrency:		        1.00
Successful transactions:         500
Failed transactions:	           0
Longest transaction:	        1.32
Shortest transaction:	        0.41

## 5ms

Run #1

Transactions:		         500 hits
Availability:		      100.00 %
Elapsed time:		      253.08 secs
Data transferred:	        3.68 MB
Response time:		        0.51 secs
Transaction rate:	        1.98 trans/sec
Throughput:		        0.01 MB/sec
Concurrency:		        1.00
Successful transactions:         500
Failed transactions:	           0
Longest transaction:	        1.82
Shortest transaction:	        0.42

Run #2

Transactions:		         500 hits
Availability:		      100.00 %
Elapsed time:		      257.79 secs
Data transferred:	        3.68 MB
Response time:		        0.52 secs
Transaction rate:	        1.94 trans/sec
Throughput:		        0.01 MB/sec
Concurrency:		        1.00
Successful transactions:         500
Failed transactions:	           0
Longest transaction:	        1.93
Shortest transaction:	        0.42
```