@rmccue 2023-02-02
Cool, that looks more like what I'd expect, no real impact to response times.