@jerico 2023-02-07
Merging despite linter check failing as there's an existing PR to fix those: https://github.com/humanmade/aws-xray/pull/79